make position and size deserialization less jank

+++ ChecklistTracker/ViewModel/LayoutParams.cs
@@ -4,19 +4,18 @@ namespace ChecklistTracker.ViewModel
 {
     public class LayoutParams
     {
-
-        public int Width { get; set; }
-        public int Height { get; set; }
+        public double Width { get; set; }
+        public double Height { get; set; }
         public Thickness Padding { get; set; }
 
-        internal LayoutParams(int width, int height, Thickness padding)
+        internal LayoutParams(double width, double height, Thickness padding)
         {
             Width = width;
             Height = height;
             Padding = padding;
         }
 
-        internal LayoutParams(int width, int height) : this(width, height, new Thickness(0))
+        internal LayoutParams(double width, double height) : this(width, height, new Thickness(0))
         {
         }
     }
